refactoring of metadata classes
